requirements.txt
# === Core Data & Config ===
numpy>=1.23
pandas>=2.0
pyyaml>=6.0

# === Machine Learning & RL ===
torch>=2.2     # 기본 CPU 버전. CUDA 사용 시 별도 설치 권장 (README 참조)
gymnasium>=0.29
scikit-learn>=1.3
tqdm>=4.66     # 학습/시뮬레이션 진행 상황(progress bar) 표시용. RL 스크립트 등에서 자주 사용되며, 용량/의존성이 가벼워 기본 포함 유지.

# === LLM & Embeddings ===
llama-cpp-python>=0.2  # 설치 시 C++ 컴파일러 필요
huggingface_hub>=0.19

# === Visualization ===
matplotlib>=3.7
